\name{helpers}
\alias{print.llama.data}
\alias{print.llama.model}
\title{
Helpers
}
\usage{
\method{print}{llama.data}(x, \dots)
\method{print}{llama.model}(x, \dots)
}
\arguments{
  \item{x}{
  the object to print.
  }
  \item{\dots}{
  ignored.
  }
}
\description{
S3 print methods.
}
\author{
Lars Kotthoff
}
